2 Valentia Road
"Well, Miss Hunter, if your mind is made up, that settles the question," said Holmes, smiling.

"But you would not advise me to refuse?"

"I confess that it is not the situation which I should like to see a sister of mine apply for."

"What is the meaning of it all, Mr. Holmes?"