46 Harvey Street
"What of the rat, then?"

Sherlock Holmes took a folded paper from his pocket and flattened it out on the table. "This is a
map of the Colony of Victoria," he said. "I wired to Bristol for it last night." He put his hand
over part of the map. "What do you read?"

"ARAT," I read.

"And now?" He raised his hand.

"BALLARAT."

"Quite so. That was the word the man uttered, and of which his son only caught the last two
syllables. He was trying to utter the name of his murderer. So and so, of Ballarat."