113 Holmefield Avenue
There was no rain, as Holmes had foretold, and the morning broke bright and cloudless. At nine
o'clock Lestrade called for us with the carriage, and we set off for Hatherley Farm and the Boscombe
Pool.

"There is serious news this morning," Lestrade observed. "It is said that Mr. Turner, of the Hall,
is so ill that his life is despaired of."

"An elderly man, I presume?" said Holmes.

"About sixty; but his constitution has been shattered by his life abroad, and he has been in failing
health for some time. This business has had a very bad effect upon him. He was an old friend of
McCarthy's, and, I may add, a great benefactor to him, for I have learned that he gave him Hatherley
Farm rent free."