15 Cliff Place
"He has, however, retained some degree of self-respect," he continued, disregarding my remonstrance.
"He is a man who leads a sedentary life, goes out little, is out of training entirely, is middle-
aged, has grizzled hair which he has had cut within the last few days, and which he anoints with
lime-cream. These are the more patent facts which are to be deduced from his hat. Also, by the way,
that it is extremely improbable that he has gas laid on in his house."

"You are certainly joking, Holmes."

"Not in the least. Is it possible that even now, when I give you these results, you are unable to
see how they are attained?"

"I have no doubt that I am very stupid, but I must confess that I am unable to follow you. For
example, how did you deduce that this man was intellectual?"