54 Holmefield Avenue
"I see. Then at the gasfitters' ball you met, as I understand, a gentleman called Mr. Hosmer Angel."

"Yes, sir. I met him that night, and he called next day to ask if we had got home all safe, and
after that we met him--that is to say, Mr. Holmes, I met him twice for walks, but after that father
came back again, and Mr. Hosmer Angel could not come to the house any more."

"No?"

"Well, you know father didn't like anything of the sort. He wouldn't have any visitors if he could
help it, and he used to say that a woman should be happy in her own family circle. But then, as I
used to say to mother, a woman wants her own circle to begin with, and I had not got mine yet."

"But how about Mr. Hosmer Angel? Did he make no attempt to see you?"

"Well, father was going off to France again in a week, and Hosmer wrote and said that it would be
safer and better not to see each other until he had gone. We could write in the meantime, and he
used to write every day. I took the letters in in the morning, so there was no need for father to
know."