19 Wellswood Gardens
"'Whatever were you doing with that bird, Jem?' says she.

"'Well,' said I, 'you said you'd give me one for Christmas, and I was feeling which was the
fattest.'

"'Oh,' says she, 'we've set yours aside for you--Jem's bird, we call it. It's the big white one over
yonder. There's twenty-six of them, which makes one for you, and one for us, and two dozen for the
market.'