102 Cliff Place
"As it happens, he spoke of coming into town to-day upon some most important business. It is
probable that he will be away all day, and that there would be nothing to disturb you. We have a
housekeeper now, but she is old and foolish, and I could easily get her out of the way."

"Excellent. You are not averse to this trip, Watson?"

"By no means."

"Then we shall both come. What are you going to do yourself?"

"I have one or two things which I would wish to do now that I am in town. But I shall return by the
twelve o'clock train, so as to be there in time for your coming."

"And you may expect us early in the afternoon. I have myself some small business matters to attend
to. Will you not wait and breakfast?"