18 Holmefield Avenue
"I am to be neutral?"

"To do nothing whatever. There will probably be some small unpleasantness. Do not join in it. It
will end in my being conveyed into the house. Four or five minutes afterwards the sitting-room
window will open. You are to station yourself close to that open window."

"Yes."