97 Sunningdale Drive
"This time, at least, I did not scorn her advice. I staggered to my feet and ran with her along the
corridor and down a winding stair. The latter led to another broad passage, and just as we reached
it we heard the sound of running feet and the shouting of two voices, one answering the other from
the floor on which  we were and from the one beneath. My guide stopped and looked about her like one
who is at her wit's end. Then she threw open a door which led into a bedroom, through the window of
which the moon was shining brightly.

"'It is your only chance,' said she. 'It is high, but it may be that you can jump it.'