14 Larch Walk
"I think that there you ask a little too much," responded his Lordship. "I may be forced to
acquiesce in these recent developments, but I can hardly be expected to make merry over them. I
think that with your permission I will now wish you all a very good-night." He included us all in a
sweeping bow and stalked out of the room.

"Then I trust that you at least will honour me with your company," said Sherlock Holmes. "It is
always a joy to meet an American, Mr. Moulton, for I am one of those who believe that the folly of a
monarch and the blundering of a minister in far-gone years will not prevent our children from being
some day citizens of the same world-wide country under a flag which shall be a quartering of the
Union Jack with the Stars and Stripes."

"The case has been an interesting one," remarked Holmes when our visitors had left us, "because it
serves to show very clearly how simple the explanation may be of an affair which at first sight
seems to be almost inexplicable. Nothing could be more natural than the sequence of events as
narrated by this lady, and nothing stranger than the result when viewed, for instance, by Mr.
Lestrade of Scotland Yard."

"You were not yourself at fault at all, then?"

"From the first, two facts were very obvious to me, the one that the lady had been quite willing to
undergo the wedding ceremony, the other that she had repented of it within a few minutes of
returning home. Obviously something had occurred during the morning, then, to cause her to change
her mind. What could that something be? She could not have spoken to anyone when she was out, for
she had been in the company of the bridegroom. Had she seen someone, then? If she had, it must be
someone from America because she had spent so short a time in this country that she could hardly
have allowed anyone to acquire so deep an influence over her that the mere sight of him would induce
her to change her plans so completely. You see we have already arrived, by a process of exclusion,
at the idea that she might have seen an American. Then who could this American be, and why should he
possess so much influence over her? It might be a lover; it might be a husband. Her young womanhood
had, I knew, been spent in rough scenes and under strange conditions. So far I had got before I ever
heard Lord St. Simon's narrative. When he told us of a man in a pew, of the change in the bride's
manner, of so transparent a device for obtaining a note as the dropping of a bouquet, of her resort
to her confidential maid, and of her very significant allusion to claim-jumping--which in miners'
parlance means taking possession of that which another person has a prior claim to--the whole
situation became absolutely clear. She had gone off with a man, and the man was either a lover or
was a previous husband--the chances being in favour of the latter."