141 Beaconside
"And Irene Adler?"

"Threatens to send them the photograph. And she will do it. I know that she will do it. You do not
know her, but she has a soul of steel. She has the face of the most beautiful of women, and the mind
of the most resolute of men. Rather than I should marry another woman, there are no lengths to which
she would not go--none."

"You are sure that she has not sent it yet?"

"I am sure."

"And why?"