37 Wantage Close
From outside came the occasional cry of a night-bird, and once at our very window a long drawn
catlike whine, which told us that the cheetah was indeed at liberty. Far away we could hear the deep
tones of the parish clock, which boomed out every quarter of an hour. How long they seemed, those
quarters! Twelve struck, and one and two and three, and still we sat waiting silently for whatever
might befall.

Suddenly there was the momentary gleam of a light up in the direction of the ventilator, which
vanished immediately, but was succeeded by a strong smell of burning oil and heated metal. Someone
in the next room had lit a dark-lantern. I heard a gentle sound of movement, and then all was silent
once more, though the smell grew stronger. For half an hour I sat with straining ears. Then suddenly
another sound became audible--a very gentle, soothing sound, like that of a small jet of steam
escaping continually from a kettle. The instant that we heard it, Holmes sprang from the bed, struck
a match, and lashed furiously with his cane at the bell-pull.

"You see it, Watson?" he yelled. "You see it?"

But I saw nothing. At the moment when Holmes struck the light I heard a low, clear whistle, but the
sudden glare flashing into my weary eyes made it impossible for me to tell what it was at which my
friend lashed so savagely. I could, however, see that his face was deadly pale and filled with
horror and loathing. He had ceased to strike and was gazing up at the ventilator when suddenly there
broke from the silence of the night the most horrible cry to which I have ever listened. It swelled
up louder and louder, a hoarse yell of pain and fear and anger all mingled in the one dreadful
shriek. They say that away down in the village, and even in the distant parsonage, that cry raised
the sleepers from their beds. It struck cold to our hearts, and I stood gazing at Holmes, and he at
me, until the last echoes of it had died away into the silence from which it rose.