110 Beaconside
"When he was gone I unlocked my bureau, made sure that my treasure was safe, and locked it again.
Then I started to go round the house to see that all was secure--a duty which I usually leave to
Mary but which I thought it well to perform myself that night. As I came down the stairs I saw Mary
herself at the side window of the hall, which she closed and fastened as I approached.

"'Tell me, dad,' said she, looking, I thought, a little disturbed, 'did you give Lucy, the maid,
leave to go out to-night?'

"'Certainly not.'