8 Down Close
"Then I have just time, with your assistance, to get clear upon the subject. Turn over those papers
and arrange the extracts in their order of time, while I take a glance as to who our client is." He
picked a red-covered volume from a line of books of reference beside the mantelpiece. "Here he is,"
said he, sitting down and flattening it out upon his knee. "'Lord Robert Walsingham de Vere St.
Simon, second son of the Duke of Balmoral.' Hum! 'Arms: Azure, three caltrops in chief over a fess
sable. Born in 1846.' He's forty-one years of age, which is mature for marriage. Was Under-Secretary
for the colonies in a late administration. The Duke, his father, was at one time Secretary for
Foreign Affairs. They inherit Plantagenet blood by direct descent, and Tudor on the distaff side.
Ha! Well, there is nothing very instructive in all this. I think that I must turn to you Watson, for
something more solid."

"I have very little difficulty in finding what I want," said I, "for the facts are quite recent, and
the matter struck me as remarkable. I feared to refer them to you, however, as I knew that you had
an inquiry on hand and that you disliked the intrusion of other matters."

"Oh, you mean the little problem of the Grosvenor Square furniture van. That is quite cleared up now
--though, indeed, it was obvious from the first. Pray give me the results of your newspaper
selections."

"Here is the first notice which I can find. It is in the personal column of the Morning Post, and
dates, as you see, some weeks back: