3 Balcombe Close
"We have at least an hour before us," he remarked, "for they can hardly take any steps until the
good pawnbroker is safely in bed. Then they will not lose a minute, for the sooner they do their
work the longer time they will have for their escape. We are at present, Doctor--as no doubt you
have divined--in the cellar of the City branch of one of the principal London banks. Mr.
Merryweather is the chairman of directors, and he will explain to you that there are reasons why the
more daring criminals of London should take a considerable interest in this cellar at present."

"It is our French gold," whispered the director. "We have had several warnings that an attempt might
be made upon it."

"Your French gold?"