112 Beaconside
"Not the Countess of Morcar's blue carbuncle!" I ejaculated.

"Precisely so. I ought to know its size and shape, seeing that I have read the advertisement about
it in The Times every day lately. It is absolutely unique, and its value can only be conjectured,
but the reward offered of £1000 is certainly not within a twentieth part of the market price."

"A thousand pounds! Great Lord of mercy!" The commissionaire plumped down into a chair and stared
from one to the other of us.

"That is the reward, and I have reason to know that there are sentimental considerations in the
background which would induce the Countess to part with half her fortune if she could but recover
the gem."

"It was lost, if I remember aright, at the Hotel Cosmopolitan," I remarked.

"Precisely so, on December 22nd, just five days ago. John Horner, a plumber, was accused of having
abstracted it from the lady's jewel-case. The evidence against him was so strong that the case has
been referred to the Assizes. I have some account of the matter here, I believe." He rummaged amid
his newspapers, glancing over the dates, until at last he smoothed one out, doubled it over, and
read the following paragraph: