40 Wantage Close
"I looked through and was conscious of two glowing eyes, and of a vague figure huddled up in the
darkness.

"'Don't be frightened,' said my employer, laughing at the start which I had given. 'It's only Carlo,
my mastiff. I call him mine, but really old Toller, my groom, is the only man who can do anything
with him. We feed him once a day, and not too much then, so that he is always as keen as mustard.
Toller lets him loose every night, and God help the trespasser whom he lays his fangs upon. For
goodness' sake don't you ever on any pretext set your foot over the threshold at night, for it's as
much as your life is worth.'

"The warning was no idle one, for two nights later I happened to look out of my bedroom window about
two o'clock in the morning. It was a beautiful moonlight night, and the lawn in front of the house
was silvered over and almost as bright as day. I was standing, rapt in the peaceful beauty of the
scene, when I was aware that something was moving under the shadow of the copper beeches. As it
emerged into the moonshine I saw what it was. It was a giant dog, as large as a calf, tawny tinted,
with hanging jowl, black muzzle, and huge projecting bones. It walked slowly across the lawn and
vanished into the shadow upon the other side. That dreadful sentinel sent a chill to my heart which
I do not think that any burglar could have done.

"And now I have a very strange experience to tell you. I had, as you know, cut off my hair in
London, and I had placed it in a great coil at the bottom of my trunk. One evening, after the child
was in bed, I began to amuse myself by examining the furniture of my room and by rearranging my own
little things. There was an old chest of drawers in the room, the two upper ones empty and open, the
lower one locked. I had filled the first two with my linen, and as I had still much to pack away I
was naturally annoyed at not having the use of the third drawer. It struck me that it might have
been fastened by a mere oversight, so I took out my bunch of keys and tried to open it. The very
first key fitted to perfection, and I drew the drawer open. There was only one thing in it, but I am
sure that you would never guess what it was. It was my coil of hair.

"I took it up and examined it. It was of the same peculiar tint, and the same thickness. But then
the impossibility of the thing obtruded itself upon me. How could my hair have been locked in the
drawer? With trembling hands I undid my trunk, turned out the contents, and drew from the bottom my
own hair. I laid the two tresses together, and I assure you that they were identical. Was it not
extraordinary? Puzzle as I would, I could make nothing at all of what it meant. I returned the
strange hair to the drawer, and I said nothing of the matter to the Rucastles as I felt that I had
put myself in the wrong by opening a drawer which they had locked.

"I am naturally observant, as you may have remarked, Mr. Holmes, and I soon had a pretty good plan
of the whole house in my head. There was one wing, however, which appeared not to be inhabited at
all. A door which faced that which led into the quarters of the Tollers opened into this suite, but
it was invariably locked. One day, however, as I ascended the stair, I met Mr. Rucastle coming out
through this door, his keys in his hand, and a look on his face which made him a very different
person to the round, jovial man to whom I was accustomed. His cheeks were red, his brow was all
crinkled with anger, and the veins stood out at his temples with passion. He locked the door and
hurried past me without a word or a look.