116 Beaconside
"I bowed, feeling as flattered as any young man would at such an address. 'May I ask who it was who
gave me so good a character?'

"'Well, perhaps it is better that I should not tell you that just at this moment. I have it from the
same source that you are both an orphan and a bachelor and are residing alone in London.'

"'That is quite correct,' I answered; 'but you will excuse me if I say that I cannot see how all
this bears upon my professional qualifications. I understand that it was on a professional matter
that you wished to speak to me?'