49 Harvey Street
"By this time the whole house was astir, for I had raised my voice in my anger. Mary was the first
to rush into my room, and, at the sight of the coronet and of Arthur's face, she read the whole
story and, with a scream, fell down senseless on the ground. I sent the house-maid for the police
and put the investigation into their hands at once. When the inspector and a constable entered the
house, Arthur, who had stood sullenly with his arms folded, asked me whether it was my intention to
charge him with theft. I answered that it had ceased to be a private matter, but had become a public
one, since the ruined coronet was national property. I was determined that the law should have its
way in everything.

"'At least,' said he, 'you will not have me arrested at once. It would be to your advantage as well
as mine if I might leave the house for five minutes.'

"'That you may get away, or perhaps that you may conceal what you have stolen,' said I. And then,
realising the dreadful position in which I was placed, I implored him to remember that not only my
honour but that of one who was far greater than I was at stake; and that he threatened to raise a
scandal which would convulse the nation. He might avert it all if he would but tell me what he had
done with the three missing stones.