51 Wantage Close
Lestrade looked startled. "I do not quite follow," he said.

"How is the glass? Twenty-nine, I see. No wind, and not a cloud in the sky. I have a caseful of
cigarettes here which need smoking, and the sofa is very much superior to the usual country hotel
abomination. I do not think that it is probable that I shall use the carriage to-night."

Lestrade laughed indulgently. "You have, no doubt, already formed your conclusions from the
newspapers," he said. "The case is as plain as a pikestaff, and the more one goes into it the
plainer it becomes. Still, of course, one can't refuse a lady, and such a very positive one, too.
She has heard of you, and would have your opinion, though I repeatedly told her that there was
nothing which you could do which I had not already done. Why, bless my soul! here is her carriage at
the door."