59 Bowler Street
"'Ample.'

"'You understand, Mr. Holder, that I am giving you a strong proof of the confidence which I have in
you, founded upon all that I have heard of you. I rely upon you not only to be discreet and to
refrain from all gossip upon the matter but, above all, to preserve this coronet with every possible
precaution because I need not say that a great public scandal would be caused if any harm were to
befall it. Any injury to it would be almost as serious as its complete loss, for there are no beryls
in the world to match these, and it would be impossible to replace them. I leave it with you,
however, with every confidence, and I shall call for it in person on Monday morning.'

"Seeing that my client was anxious to leave, I said no more but, calling for my cashier, I ordered
him to pay over fifty £1000 notes. When I was alone once more, however, with the precious case lying
upon the table in front of me, I could not but think with some misgivings of the immense
responsibility which it entailed upon me. There could be no doubt that, as it was a national
possession, a horrible scandal would ensue if any misfortune should occur to it. I already regretted
having ever consented to take charge of it. However, it was too late to alter the matter now, so I
locked it up in my private safe and turned once more to my work.

"When evening came I felt that it would be an imprudence to leave so precious a thing in the office
behind me. Bankers' safes had been forced before now, and why should not mine be? If so, how
terrible would be the position in which I should find myself! I determined, therefore, that for the
next few days I would always carry the case backward and forward with me, so that it might never be
really out of my reach. With this intention, I called a cab and drove out to my house at Streatham,
carrying the jewel with me. I did not breathe freely until I had taken it upstairs and locked it in
the bureau of my dressing-room.

"And now a word as to my household, Mr. Holmes, for I wish you to thoroughly understand the
situation. My groom and my page sleep out of the house, and may be set aside altogether. I have
three maid-servants who have been with me a number of years and whose absolute reliability is quite
above suspicion. Another, Lucy Parr, the second waiting-maid, has only been in my service a few
months. She came with an excellent character, however, and has always given me satisfaction. She is
a very pretty girl and has attracted admirers who have occasionally hung about the place. That is
the only drawback which we have found to her, but we believe her to be a thoroughly good girl in
every way.

"So much for the servants. My family itself is so small that it will not take me long to describe
it. I am a widower and have an only son, Arthur. He has been a disappointment to me, Mr. Holmes--a
grievous disappointment. I have no doubt that I am myself to blame. People tell me that I have
spoiled him. Very likely I have. When my dear wife died I felt that he was all I had to love. I
could not bear to see the smile fade even for a moment from his face. I have never denied him a
wish. Perhaps it would have been better for both of us had I been sterner, but I meant it for the
best.