5 Corbet Close
"I thought it as well," said Holmes as we climbed the stile, "that this fellow should think we had
come here as architects, or on some definite business. It may stop his gossip. Good-afternoon, Miss
Stoner. You see that we have been as good as our word."

Our client of the morning had hurried forward to meet us with a face which spoke her joy. "I have
been waiting so eagerly for you," she cried, shaking hands with us warmly. "All has turned out
splendidly. Dr. Roylott has gone to town, and it is unlikely that he will be back before evening."