56 Valentia Road
Fairbank was a good-sized square house of white stone, standing back a little from the road. A
double carriage-sweep, with a snow-clad lawn, stretched down in front to two large iron gates which
closed the entrance. On the right side was a small wooden thicket, which led into a narrow path
between two neat hedges stretching from the road to the kitchen door, and forming the tradesmen's
entrance. On the left ran a lane which led to the stables, and was not itself within the grounds at
all, being a public, though little used, thoroughfare. Holmes left us standing at the door and
walked slowly all round the house, across the front, down the tradesmen's path, and so round by the
garden behind into the stable lane. So long was he that Mr. Holder and I went into the dining-room
and waited by the fire until he should return. We were sitting there in silence when the door opened
and a young lady came in. She was rather above the middle height, slim, with dark hair and eyes,
which seemed the darker against the absolute pallor of her skin. I do not think that I have ever
seen such deadly paleness in a woman's face. Her lips, too, were bloodless, but her eyes were
flushed with crying. As she swept silently into the room she impressed me with a greater sense of
grief than the banker had done in the morning, and it was the more striking in her as she was
evidently a woman of strong character, with immense capacity for self-restraint. Disregarding my
presence, she went straight to her uncle and passed her hand over his head with a sweet womanly
caress.

"You have given orders that Arthur should be liberated, have you not, dad?" she asked.

"No, no, my girl, the matter must be probed to the bottom."